Backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1

langchain==0.2.12
langchain-openai==0.1.21
openai>=1.30.0
tiktoken>=0.7.0

faiss-cpu==1.7.4
numpy<2.0.0
